requirements.txt
fastapi
uvicorn
numpy
pillow
scikit-learn
